doc: note that $(wildcard) is implemented in terms of glob(3)
The $subject isn't strictly true, as the it's implemented in terms of
MatchFirst() on Amiga, but that's obscure enough to not be noted in
the manual, and in any case the tests added in 0d366b66 (* Added the
test suite to the main distribution., 1999-09-14) show that the output
is sorted on all platforms GNU make has been tested on.

The motivation for this patch is that I've seen code like this in the
wild:

    FILES = $(sort $(wildcard t*.sh))

I thought that it would be documented that that sort of thing was
redundant, but I didn't find any mention of it in the manual. After
some source spelunking I came up with this addition, which would have
helped me when I consulted the manual.